+++ apps/client/assets/js/routes/ijust/IjustContextEventRoute.tsx
@@ -1,28 +1,17 @@
 import React, { useCallback, useState } from "react";
 import { format, formatDistanceToNow, parseISO } from "date-fns";
 import { useParams } from "react-router-dom";
+import Breadcrumbs from "@mui/material/Breadcrumbs";
+import Button from "@mui/material/Button";
+import { Link } from "react-router-dom";
 
 import { MainNav } from "./../../components/MainNav";
 import { IjustEventOccurrences } from "./../../components/ijust/IjustEventOccurrences";
 import { Constants } from "./../../utils/Constants";
 import { QueryLoader } from "./../../utils/QueryLoader";
 import { graphql } from "../../gql";
+import { IjustEditEventModal } from "./../../components/ijust/IjustEventEditModal";
 import type { IjustEvent, IjustContext } from "@gql-types";
-
-import Breadcrumbs from "@mui/material/Breadcrumbs";
-import Button from "@mui/material/Button";
-import { Link } from "react-router-dom";
-import { useMutation } from "urql";
-import * as yup from "yup";
-import { Controller, useForm } from "react-hook-form";
-import { yupResolver } from "@hookform/resolvers/yup";
-
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogTitle from "@mui/material/DialogTitle";
-import TextField from "@mui/material/TextField";
-import Alert from "@mui/material/Alert";
 
 const QUERY = graphql(`
   query GetEvent($contextId: ID!, $eventId: ID!) {
@@ -40,26 +29,6 @@ const QUERY = graphql(`
       ijustContext {
         id
         name
-      }
-    }
-  }
-`);
-
-const UPDATE_EVENT = graphql(`
-  mutation UpdateIjustEvent($id: ID!, $name: String, $cost: Int) {
-    updateIjustEvent(id: $id, name: $name, cost: $cost) {
-      successful
-      messages {
-        message
-        field
-      }
-      result {
-        id
-        name
-        cost {
-          amount
-          currency
-        }
       }
     }
   }
@@ -87,11 +56,12 @@ export function IjustContextEventRoute() {
 }
 
 interface IjustEventComponentProps {
-  event: IjustEvent
-  context: IjustContext
+  event: IjustEvent;
+  context: IjustContext;
 }
 
 function IjustEventComponent({ event, context }: IjustEventComponentProps) {
+  const [editing, setEditing] = useState(false);
 
   return (
     <div>
@@ -102,151 +72,24 @@ function IjustEventComponent({ event, context }: IjustEventComponentProps) {
           <h1 className="text-xl">{event.name}</h1>
         </Breadcrumbs>
       </div>
-      <Header event={event} />
+      <IjustEditEventModal
+        event={event}
+        visible={editing}
+        setVisible={setEditing}
+      />
+      <Button onClick={() => setEditing(true)}>Edit</Button>
+      <EventInfo event={event} />
       <IjustEventOccurrences contextId={context.id} eventId={event.id} />
     </div>
-  )
-};
-
-interface FormInputs {
-  cost: number;
-  name: string;
-  backendError: null;
+  );
 }
 
-const schema = yup
-  .object({
-    cost: yup.number().positive().transform((_, value) => {
-      return value === "" ? null : Number(value);
-    }).nullable(),
-    name: yup.string().required(),
-  })
-  .required();
-
-interface HeaderProps {
-  event: IjustEvent
+interface EventInfoProps {
+  event: IjustEvent;
 }
-function Header({ event }: HeaderProps) {
-  const [editing, setEditing] = useState(false);
-  const [_result, updateEvent] = useMutation(UPDATE_EVENT);
-
-  const {
-    clearErrors,
-    control,
-    formState: { isValid, errors, isSubmitting },
-    handleSubmit,
-    reset,
-    setError,
-  } = useForm<FormInputs>({
-    defaultValues: {
-      cost: event.cost?.amount || "",
-      name: event.name,
-      backendError: null,
-    },
-    mode: "onBlur",
-    resolver: yupResolver(schema),
-  });
-
-  const setBackendError = useCallback(
-    (message: string) => {
-      setError("backendError", { type: "custom", message });
-    },
-    [setError]
-  );
-
-  const handleClose = () => {
-    reset();
-    setEditing(false);
-  }
-
-  const onSubmit = useCallback(
-    async (form: FormInputs) => {
-      clearErrors("backendError");
-      const { cost } = form;
-
-      try {
-        const { error, data } = await updateEvent({ id: event.id, cost, name });
-        console.log(error);
-        console.log(data);
-        debugger;
-
-        if (error) {
-          console.error(error);
-          setBackendError(error.message);
-        } else {
-          reset();
-          return;
-        }
-      } catch (e) {
-        console.error(e);
-        setBackendError("Something went wrong. Please try again.");
-      }
-    },
-    [clearErrors, setError, updateEvent]
-  );
-
+function EventInfo({ event }: EventInfoProps) {
   return (
     <>
-      <Dialog
-        open={editing}
-        onClose={handleClose}
-        aria-labelledby="edit-event-title"
-      >
-        <DialogTitle id="edit-event-title">Edit event</DialogTitle>
-        <form onSubmit={handleSubmit(onSubmit)}>
-          <DialogContent>
-            {errors.backendError?.message && (
-              <Alert color="error">{errors.backendError.message}</Alert>
-            )}
-            <Controller
-              control={control}
-              name="name"
-              render={({ field }) => (
-                <TextField
-                  {...field}
-                  className="mt-3"
-                  label="Name"
-                  error={!!errors.name?.message}
-                  fullWidth
-                />
-              )}
-            />
-            {errors.name?.message && (
-              <Alert color="error">{errors.name.message}</Alert>
-            )}
-
-            <Controller
-              control={control}
-              name="cost"
-              render={({ field }) => (
-                <TextField
-                  {...field}
-                  className="mt-3"
-                  type="number"
-                  label="Cost"
-                  error={!!errors.cost?.message}
-                  fullWidth
-                />
-              )}
-            />
-            {errors.cost?.message && (
-              <Alert color="error">{errors.cost.message}</Alert>
-            )}
-
-          </DialogContent>
-          <DialogActions>
-            <Button color="secondary" onClick={handleClose}>
-              Cancel
-            </Button>
-            <Button type="submit" disabled={!isValid && !isSubmitting}>
-              Save
-            </Button>
-          </DialogActions>
-        </form>
-      </Dialog>
-
-      <Button onClick={() => setEditing(true)}>Edit</Button>
-
       <table className="mt-3 w-full">
         <tbody>
           <tr>
@@ -256,7 +99,10 @@ function Header({ event }: HeaderProps) {
           <tr>
             <td className="py-3">First occurred</td>
             <td className="py-3">
-              {format(parseISO(event.insertedAt + "Z"), Constants.dateTimeFormat)}
+              {format(
+                parseISO(event.insertedAt + "Z"),
+                Constants.dateTimeFormat
+              )}
               <span className="ml-3">
                 ({formatDistanceToNow(parseISO(event.insertedAt + "Z"))} ago)
               </span>
@@ -265,7 +111,10 @@ function Header({ event }: HeaderProps) {
           <tr>
             <td className="py-3">Last occurred</td>
             <td className="py-3">
-              {format(parseISO(event.updatedAt + "Z"), Constants.dateTimeFormat)}
+              {format(
+                parseISO(event.updatedAt + "Z"),
+                Constants.dateTimeFormat
+              )}
               <span className="ml-3">
                 ({formatDistanceToNow(parseISO(event.updatedAt + "Z"))} ago)
               </span>
@@ -282,17 +131,17 @@ function Header({ event }: HeaderProps) {
 }
 
 interface Money {
-  amount: number
-  currency: string
+  amount: number;
+  currency: string;
 }
 
 function formatMoney(money?: Money) {
   if (!money) return "-";
 
-  const formatter = new Intl.NumberFormat(
-    "en-US",
-    { style: "currency", currency: money.currency }
-  );
+  const formatter = new Intl.NumberFormat("en-US", {
+    style: "currency",
+    currency: money.currency,
+  });
 
   return formatter.format(money.amount);
 }
